LayoutTests/editing/spelling/spellcheck-api.html

<html>
<head>
<title></title>
<link rel="stylesheet" href="../../fast/js/resources/js-test-style.css">
<script language="javascript" type="text/javascript" src="../../fast/js/resources/js-test-pre.js"></script>
<script language="javascript" type="text/javascript">
// The node used by SpellCheck API.
var node;

// The array of expected suggestions.
var suggestions = ['welcome'];

// The main function of this test. We can add or remove spellcheck ranges after
// we finish loading this page. Therefore, this function should be an event
// handler for onload events.
function RunTest() {
    for (var i = 0; i < 3; i++) {
        node = document.getElementById('test' + i);

        // Add a misspelled range and verify we can read the range.
        node.addSpellcheckRange(0, 8, suggestions, 0);
        shouldBeNonNull("node.spellcheckRanges");
        shouldBe("node.spellcheckRanges.length", "1");
        shouldBe("node.spellcheckRanges[0].start", "0");
        shouldBe("node.spellcheckRanges[0].length", "8");
        shouldBe("node.spellcheckRanges[0].options", "0");
        shouldBe("node.spellcheckRanges[0].suggestions.length", "1");
        shouldBe("node.spellcheckRanges[0].suggestions[0]", "suggestions[0]");

        // Remove the misspelled range and verify we have removed the range.
        node.removeSpellcheckRange(node.spellcheckRanges[0]);
        shouldBe("node.spellcheckRanges.length", "0");
    }
}
</script>
</head>
<body onload="RunTest()">
<p id="description"></p>
<div id="console"></div>
<input id="test0" type="text" value="wellcome"/>
<textarea id="test1" rows="80" cols="80">wellcome</textarea>
<div id="test2" contenteditable="true">wellcome</div>
<script language="javascript" type="text/javascript">
description('This tests SpellCheck API can mark a word as misspelled.');

var successfullyParsed = true;
</script>
<script language="javascript" type="text/javascript" src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
