LayoutTests/editing/spelling/spellcheck-api-expected.txt

This tests SpellCheck API can mark a word as misspelled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS node.spellcheckRanges is non-null.
PASS node.spellcheckRanges.length is 1
PASS node.spellcheckRanges[0].start is 0
PASS node.spellcheckRanges[0].length is 8
PASS node.spellcheckRanges[0].options is 0
PASS node.spellcheckRanges[0].suggestions.length is 1
PASS node.spellcheckRanges[0].suggestions[0] is suggestions[0]
PASS node.spellcheckRanges.length is 0
PASS node.spellcheckRanges is non-null.
PASS node.spellcheckRanges.length is 1
PASS node.spellcheckRanges[0].start is 0
PASS node.spellcheckRanges[0].length is 8
PASS node.spellcheckRanges[0].options is 0
PASS node.spellcheckRanges[0].suggestions.length is 1
PASS node.spellcheckRanges[0].suggestions[0] is suggestions[0]
PASS node.spellcheckRanges.length is 0
PASS node.spellcheckRanges is non-null.
PASS node.spellcheckRanges.length is 1
PASS node.spellcheckRanges[0].start is 0
PASS node.spellcheckRanges[0].length is 8
PASS node.spellcheckRanges[0].options is 0
PASS node.spellcheckRanges[0].suggestions.length is 1
PASS node.spellcheckRanges[0].suggestions[0] is suggestions[0]
PASS node.spellcheckRanges.length is 0
 
wellcome
